refactor: organized the different ds tests into folders
